requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
django-guardian==2.4.0
django-audit-log==0.7.0
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-storages==1.14.2
boto3==1.34.11
django-redis==5.4.0
celery[sqs]==5.3.6
sentry-sdk==1.39.1
django-filter==23.5
fhir.resources==7.1.0
Pillow==10.2.0
hl7apy==1.3.4
pydicom==2.4.4
gunicorn==21.2.0